fix all

+++ part2.py
@@ -9,4 +9,4 @@ C = (F - 32) * 5/9
 """
 
 def celsius(f):  # do not change this line
-  return (f-32)*5/9
+  return int((f-32)*5/9)
